requirements.txt
flask
requests
matplotlib.pyplot